# tests/openapi.yml
openapi: "3.0.1"
info:
  title: Pulse API
  version: "1.0"
servers:
  - url: http://localhost:8080/api
paths:
...
  /countries:
    get:
      summary: Получить список стран
      description: |
        Получение списка стран с возможной фильтрацией.

        Используется на странице регистрации для предоставления возможности выбора страны, к которой относится пользователь.
        Если хотя бы один переданный регион является некорректным, весь запрос считается некорректным.

        Если никакие из фильтров не переданы, необходимо вернуть все страны.
      operationId: listCountries
      parameters:
        - name: region
          description: |
            Возвращаемые страны должны относиться только к тем регионам, которые переданы в данном списке.

            Если передан пустой список, считайте, что фильтр по региону отсутствует.
          in: query
          schema:
            type: array
            items:
              $ref: "#/components/schemas/countryRegion"
            example:
              - Europe
              - Asia
      responses:
        "200":
          description: Список стран, соответствующих указанному фильтру. Страны должны быть отсортированы лексикографически по двухбуквенному коду.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/country"
        "400":
          description: Формат входного запроса не соответствует формату либо переданы неверные значения.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
  /countries/{alpha2}:
    get:
      summary: Получить страну по alpha2 коду
      description: |
        Получение одной страны по её уникальному двухбуквенному коду.

        Используется для получения информации по определенной стране.
      operationId: getCountry
      parameters:
        - name: alpha2
          description: |
            Возвращаемая страна должна иметь указанный alpha2 код.
          required: true
          in: path
          schema:
            $ref: "#/components/schemas/countryAlpha2"
      responses:
        "200":
          description: Страна, найденная по указанному коду.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/country"
        "404":
          description: Страна с указанным кодом не найдена.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
  /auth/register:
    post:
      summary: Регистрация нового пользователя
      description: |
        Используется для регистрации нового пользователя по логину и паролю.
      operationId: authRegister
      requestBody:
        description: Данные для регистрации пользователя.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  $ref: "#/components/schemas/userLogin"
                email:
                  $ref: "#/components/schemas/userEmail"
                password:
                  $ref: "#/components/schemas/userPassword"
                countryCode:
                  $ref: "#/components/schemas/countryAlpha2"
                isPublic:
                  $ref: "#/components/schemas/userIsPublic"
                phone:
                  $ref: "#/components/schemas/userPhone"
                image:
                  $ref: "#/components/schemas/userImage"
              required:
                - login
                - email
                - password
                - countryCode
                - isPublic
      responses:
        "201":
          description: В случае успеха возвращается профиль зарегистрированного пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    $ref: "#/components/schemas/userProfile"
                required:
                  - profile
        "400":
          description: |
            Регистрационные данные не соответствуют ожидаемому формату и требованиям. 

            Например, данную ошибку необходимо возвращать в следующих ситуациях (это не полный список):

            - Недостаточно "надежный" пароль.
            - Страна с указанным кодом не найдена.
            - Длина ссылки на аватар пользователя превышает допустимый лимит.

            Для ознакомления с форматом и требованиями к регистрационным данным обратите внимание на описание моделей в Open API спецификации.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "409":
          description: |
            Нарушено требование на уникальность авторизационных данных пользователей.

            Данный код ответа должен использоваться, если пользователь с таким e-mail, номером телефона или логином уже зарегистрирован.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
  /auth/sign-in:
    post:
      summary: Аутентификация для получения токена
      description: |
        Процедура аутентификации по логину и паролю позволяет получить токен, который в дальнейшем будет использоваться пользователем для выполнения операций, требующих авторизацию.

        Сервер должен генерировать уникальные токены, имеющие время жизни (на усмотрение разработчика, от 1 до 24 часов). После истечения времени действия токен должен быть недействительным и не может использоваться для аутентификации.

        Токен является уникальным строковым значением с высокой энтропией (злоумышленник не сможет его "подобрать" перебором). При каждой новой аутентификации генерируется новый уникальный токен, который ранее не был использован. Можно использовать JWT.

        В дальнейшем полученный токен будет использоваться для авторизации пользовательских запросов. Значение токена будет подставляться в заголовок `Authorization` в формате `Bearer {token}`. Следовательно, сервер должен уметь идентифицировать пользователя по токену.
      operationId: authSignIn
      requestBody:
        description: Данные для аутентификации пользователя.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  $ref: "#/components/schemas/userLogin"
                password:
                  $ref: "#/components/schemas/userPassword"
              required:
                - login
                - password
      responses:
        "200":
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Сгенерированный токен пользователя
                    minLength: 20
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
                required:
                  - token
        "401":
          description: Пользователь с указанным логином и паролем не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
...
components:
  schemas:
    countryAlpha2:
      type: string
      description: Двухбуквенный код, уникально идентифицирующий страну
      maxLength: 2
      pattern: "[a-zA-Z]{2}"
      example: RU
    countryRegion:
      type: string
      description: Географический регион, к которому относится страна
      enum:
        - Europe
        - Africa
        - Americas
        - Oceania
        - Asia
    country:
      type: object
      description: Информация о стране из стандарта ISO 3166
      properties:
        name:
          type: string
          description: Полное название страны
          maxLength: 100
        alpha2:
          $ref: "#/components/schemas/countryAlpha2"
        alpha3:
          type: string
          description: Трехбуквенный код страны
          maxLength: 3
          pattern: "[a-zA-Z]{3}"
        region:
          $ref: "#/components/schemas/countryRegion"
      required:
        - name
        - alpha2
        - alpha3
      example:
        name: Burkina Faso
        alpha2: BF
        alpha3: BFA
        region: Africa
    userLogin:
      type: string
      description: Логин пользователя
      maxLength: 30
      pattern: "[a-zA-Z0-9-]+"
      example: yellowMonkey
    userEmail:
      type: string
      description: E-mail пользователя
      maxLength: 50
      minLength: 1
      example: yellowstone1980@you.ru
    userPassword:
      type: string
      description: |
        Пароль пользователя, к которому предъявляются следующие требования:

        - Длина пароля не менее 6 символов.
        - Присутствуют латинские символы в нижнем и верхнем регистре.
        - Присутствует минимум одна цифра.
      minLength: 6
      maxLength: 100
      example: $aba4821FWfew01#.fewA$
    userIsPublic:
      type: boolean
      description: |
        Является ли данный профиль публичным. 

        Публичные профили доступны другим пользователям: если профиль публичный, любой пользователь платформы сможет получить информацию о пользователе.
      example: true
    userPhone:
      type: string
      description: Номер телефона пользователя в формате +123456789
      pattern: \+[\d]+
      example: "+74951239922"
      maxLength: 20
    userImage:
      type: string
      description: Ссылка на фото для аватара пользователя
      example: https://http.cat/images/100.jpg
      maxLength: 200
      minLength: 1
    userProfile:
      type: object
      description: Информация о профиле пользователя
      properties:
        login:
          $ref: "#/components/schemas/userLogin"
        email:
          $ref: "#/components/schemas/userEmail"
        countryCode:
          $ref: "#/components/schemas/countryAlpha2"
        isPublic:
          $ref: "#/components/schemas/userIsPublic"
        phone:
          $ref: "#/components/schemas/userPhone"
        image:
          $ref: "#/components/schemas/userImage"
      required:
        - login
        - email
        - countryCode
        - isPublic
...
    errorResponse:
      type: object
      description: Используется для возвращения ошибки пользователю
      properties:
        reason:
          type: string
          description: Описание ошибки в свободной форме
          minLength: 5
      required:
        - reason
      example:
        reason: <объяснение, почему запрос пользователя не может быть обработан>
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer